+++ examples/hello-world-app/src/views/pages/home/main.js
@@ -1,4 +1,4 @@
-import App from 'App/app';
+import App from 'App/App.js';
 
 window.addEventListener('load', function (event) {
   console.log('Start app!');