+++ src/Modules.js
@@ -1,11 +1,11 @@
-// Proxy modules available in both pug-plugin and pug-loader.
+// Modules available in both pug-plugin and pug-loader.
 // These modules store data for exchange between the plugin and the loader.
 
 const pugLoaderPath = '@webdiscus/pug-loader';
 
 // For local development of pug-plugin and pug-loader only.
-// Update path in both pug-loader/src/ModuleProxy.js and pug-plugin/src/ModuleProxy.js files.
-//const pugLoaderPath = '../';
+// Update path in both pug-loader/src/Modules.js and pug-plugin/src/Modules.js files.
+//const pugLoaderPath = '..';
 
 module.exports = {
   plugin: require(pugLoaderPath + '/src/Plugin.js'),