+++ test/utils/webpack.js
@@ -3,7 +3,7 @@ const path = require('path');
 const webpack = require('webpack');
 const { merge } = require('webpack-merge');
 
-const { plugin } = require('../../src/ModuleProxy');
+const { plugin } = require('../../src/Modules');
 const htmlWebpackPlugin = require('../../src/extras/HtmlWebpackPlugin');
 
 const prepareWebpackConfig = (PATHS, relTestCasePath, webpackOpts = {}) => {