+++ test/webpack.common.js
@@ -17,10 +17,5 @@ module.exports = {
     rules: [],
   },
 
-  optimization: {
-    // removeEmptyChunks: true,
-    // mergeDuplicateChunks: true,
-    // usedExports: true,
-    // concatenateModules: true,
-  },
+  optimization: {},
 };